requirements.txt
dotmap==1.3.17
smbus2~=0.3.0
Flask==1.1.2
Jinja2==2.10.1
numpy==1.17.4
scipy==1.3.3
matplotlib==3.2.0
toml==0.10.0
pyserial==3.4
# canlib==1.12.251
pyvisa==1.10.1
PyVisa-py==0.4.1


construct~=2.10.56
canlib~=1.12.251
# reikna~=0.7.5
v4l2~=0.2